Rename router file
